+++ ios/InteractiveReader/InteractiveReader/Features/Playback/JobPlaybackView.swift
@@ -40,7 +40,8 @@ struct JobPlaybackView: View {
     @State var resumeDecisionPending = false
     @State var pendingInteractiveAutoplayID: UUID?
     @State var nowPlayingReassertionTask: Task<Void, Never>?
-    @State var lastReaderTransportToggleTime: TimeInterval = 0
+    @State var lastReaderTransportCommandTime: TimeInterval = 0
+    @State var lastReaderTransportAction = "none"
     #if DEBUG
     @State var e2eReaderTransportCommandCount = 0
     @State var e2eTVPlayPauseCommandCount = 0
@@ -110,7 +111,7 @@ struct JobPlaybackView: View {
         e2eTVPlayPauseCommandCount += 1
         #endif
         playbackLogger.info("Job foreground tvOS Play/Pause command")
-        toggleReaderNowPlayingTransport()
+        toggleReaderNowPlayingTransport(source: "foreground")
     }
     #endif
 
@@ -448,7 +449,8 @@ struct JobPlaybackView: View {
                 musicOwnership: musicOwnership,
                 audioCoordinator: viewModel.audioCoordinator,
                 readerTransportCommandCount: e2eReaderTransportCommandCount,
-                foregroundPlayPauseCount: e2eTVPlayPauseCommandCount
+                foregroundPlayPauseCount: e2eTVPlayPauseCommandCount,
+                lastReaderTransportAction: lastReaderTransportAction
             )
         }
         #endif